+++ src/common.js
@@ -1,1 +1,3 @@
 const TYPES = { LOCAL: "local", SESSION: "session", COOKIE: "cookie" };
+
+export { TYPES };
