+++ src/page.js
@@ -1,4 +1,8 @@
-function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
+import { TYPES } from "./common.js";
+import { fireEvent } from "./handler/analytics.js";
+import { newSessionItem, newLocalItem, newCookieItem } from "./view/items.js";
+
+function appPage({ storeSave, storeDelete, storeSet, refreshValues }) {
   const PAGES = {
     EMPTY: {
       emptyPage: true,
@@ -36,7 +40,7 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
     if (formValidation()) onSaveItem();
     else {
       const invalidFields = [...document.querySelectorAll(".form-control:invalid")].map((element) => element.name);
-      analytics.fireEvent("invalid_form", { invalidFields: invalidFields.toString() });
+      fireEvent("invalid_form", { invalidFields: invalidFields.toString() });
     }
   });
 
@@ -104,13 +108,13 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
       let item;
       switch (presentationItem.type) {
         case TYPES.SESSION:
-          item = creator.newSessionItem(presentationItem, actions);
+          item = newSessionItem(presentationItem, actions);
           break;
         case TYPES.LOCAL:
-          item = creator.newLocalItem(presentationItem, actions);
+          item = newLocalItem(presentationItem, actions);
           break;
         case TYPES.COOKIE:
-          item = creator.newCookieItem(presentationItem, actions);
+          item = newCookieItem(presentationItem, actions);
       }
       keyListElement.appendChild(item);
     });
@@ -166,7 +170,7 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
   }
 
   function onDeleteItem(itemId) {
-    analytics.fireEvent("delete_item");
+    fireEvent("delete_item");
     const idToRemove = presentationItems.find((element) => itemId === element.id).id;
 
     storeDelete(idToRemove);
@@ -175,7 +179,7 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
   async function onSetItemValue(itemId, value) {
     const item = presentationItems.find(({ id }) => id === itemId);
     await storeSet(item, value);
-    analytics.fireEvent("set_value");
+    fireEvent("set_value");
 
     const viewBtn = document.getElementById(`viewBtn-${itemId}`);
     onViewValue(itemId, viewBtn);
@@ -183,7 +187,7 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
 
   function onCopyValue(element, itemId) {
     const token = document.getElementById(`token-${itemId}`)?.innerHTML;
-    analytics.fireEvent("copy_value");
+    fireEvent("copy_value");
     navigator.clipboard.writeText(token).then(
       function () {
         element.lastChild.classList.remove("fa-copy");
@@ -202,7 +206,7 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
   }
 
   function onViewValue(itemId, element) {
-    analytics.fireEvent("view_value");
+    fireEvent("view_value");
     const textArea = document.getElementById(`token-${itemId}`);
     textArea.disabled = true;
 
@@ -259,3 +263,5 @@ function appPage(creator, {storeSave, storeDelete, storeSet, refreshValues}) {
     PAGES,
   };
 }
+
+export default appPage;
