+++ src/popup.html
@@ -111,21 +111,7 @@
         </button>
       </div>
     </footer>
-    <script src="common.js"></script>
-    <script src="./handler/analytics.js"></script>
-    <script src="./handler/operation/remote.js"></script>
-    <script src="./handler/operation/cookie.js"></script>
-    <script src="./handler/valueReader.js"></script>
-    <script src="./handler/valueWriter.js"></script>
-    <script src="./handler/store.js"></script>
-    <script src="./handler/versionController.js"></script>
-
-    <script src="manager.js"></script>
-    <script src="components.js"></script>
-    <script src="items.js"></script>
-    <script src="page.js"></script>
-
-    <script src="popup.js"></script>
+    <script type="module" src="popup.js"></script>
     <script src="./assets/bootstrap/js/bootstrap.min.js"></script>
   </body>
 </html>
